fieldmanagertest: Reduce API surface of the test package
